Create TrackContainer
This PR creates the class which represent the space in the
CaptureWindows where Tracks will be drawn. See
go/stadia-orbit-timeline-improvements for more information.

Apart of a few minor changes, this PR is mostly moving function from
TimeGraph to the new TrackContainer class. Basically, TrackContainer
will manage all related with the vertical space and zooming that
previously was in the TimeGraph and all the logic related with Tracks.

Horizontal changes which affects the visible time are still managed
inside TimeGraph for now.

There are a few functions that have to be moved around:

 - ProcessTimer functions shouldn't be inside the UI, and Track creation
should be managed by previously querying CaptureDataProvider by
TrackManager. This is not a priority now, and meanwhile we can move
these functions to TrackManager, but is not a focus of this PR.
(http://b/214282122).

- SortedTracks should be managed inside TrackContainer instead of
TrackManager. This needs a few modifications that aren't either a
priority (http://b/214280810).

TODO:
1- I'm probably breaking E2E tests here. I will look at it in the
following days.
2- Next task is making TimelineUI a child of TimeGraph and
erasing the old timeline from CaptureWindows.

TEST: Load a capture. Start a capture, test some basic functionalities.

Bug: http://b/208446487
