+++ src/OrbitGl/TimeGraph.h
@@ -5,21 +5,14 @@
 #ifndef ORBIT_GL_TIME_GRAPH_H_
 #define ORBIT_GL_TIME_GRAPH_H_
 
-#include <absl/container/flat_hash_map.h>
-
 #include <cstdint>
-#include <map>
 #include <memory>
-#include <optional>
-#include <string>
 #include <vector>
 
 #include "AccessibleInterfaceProvider.h"
 #include "Batcher.h"
-#include "CallstackThreadBar.h"
 #include "CaptureViewElement.h"
 #include "ClientData/CaptureData.h"
-#include "ClientData/TimerChain.h"
 #include "ClientProtos/capture_data.pb.h"
 #include "CoreMath.h"
 #include "ManualInstrumentationManager.h"
@@ -28,10 +21,8 @@
 #include "TextRenderer.h"
 #include "TimeGraphLayout.h"
 #include "TimelineInfoInterface.h"
-#include "Track.h"
-#include "TrackManager.h"
+#include "TrackContainer.h"
 #include "Viewport.h"
-#include "absl/container/flat_hash_map.h"
 
 class OrbitApp;
 
@@ -59,9 +50,13 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
   [[nodiscard]] const orbit_client_data::CaptureData* GetCaptureData() const {
     return capture_data_;
   }
-  [[nodiscard]] orbit_gl::TrackManager* GetTrackManager() { return track_manager_.get(); }
+  [[nodiscard]] orbit_gl::TrackContainer* GetTrackContainer() const {
+    return track_container_.get();
+  }
+  [[nodiscard]] orbit_gl::TrackManager* GetTrackManager() const {
+    return track_container_->GetTrackManager();
+  }
 
-  [[nodiscard]] float GetTextBoxHeight() const { return layout_.GetTextBoxHeight(); }
   [[nodiscard]] float GetWorldFromTick(uint64_t time) const override;
   [[nodiscard]] float GetWorldFromUs(double micros) const override;
   [[nodiscard]] uint64_t GetTickFromWorld(float world_x) const override;
@@ -90,8 +85,6 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
   void HorizontallyMoveIntoView(VisibilityType vis_type,
                                 const orbit_client_protos::TimerInfo& timer_info,
                                 double distance = 0.3);
-  void VerticallyMoveIntoView(const orbit_client_protos::TimerInfo& timer_info);
-  void VerticallyMoveIntoView(Track& track);
 
   [[nodiscard]] double GetTime(double ratio) const;
   void SelectAndMakeVisible(const orbit_client_protos::TimerInfo* timer_info);
@@ -99,20 +92,11 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
   enum class JumpDirection { kPrevious, kNext, kTop, kDown };
   void JumpToNeighborTimer(const orbit_client_protos::TimerInfo* from, JumpDirection jump_direction,
                            JumpScope jump_scope);
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindPreviousFunctionCall(
-      uint64_t function_address, uint64_t current_time,
-      std::optional<uint32_t> thread_id = std::nullopt) const;
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindNextFunctionCall(
-      uint64_t function_address, uint64_t current_time,
-      std::optional<uint32_t> thread_id = std::nullopt) const;
-  [[nodiscard]] std::vector<const orbit_client_protos::TimerInfo*> GetAllTimersForHookedFunction(
-      uint64_t function_address) const;
 
   void SelectAndZoom(const orbit_client_protos::TimerInfo* timer_info);
   [[nodiscard]] double GetCaptureTimeSpanUs() const;
   [[nodiscard]] double GetCurrentTimeSpanUs() const;
   [[nodiscard]] bool IsRedrawNeeded() const { return update_primitives_requested_; }
-  void SetThreadFilter(const std::string& filter);
 
   [[nodiscard]] bool IsFullyVisible(uint64_t min, uint64_t max) const;
   [[nodiscard]] bool IsPartlyVisible(uint64_t min, uint64_t max) const;
@@ -120,25 +104,10 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
 
   [[nodiscard]] TextRenderer* GetTextRenderer() { return &text_renderer_static_; }
   [[nodiscard]] Batcher& GetBatcher() { return batcher_; }
-  [[nodiscard]] std::vector<const orbit_client_data::TimerChain*> GetAllThreadTrackTimerChains()
-      const;
-  [[nodiscard]] int GetNumVisiblePrimitives() const;
 
   void UpdateHorizontalScroll(float ratio);
   [[nodiscard]] const TimeGraphLayout& GetLayout() const { return layout_; }
   [[nodiscard]] TimeGraphLayout& GetLayout() { return layout_; }
-
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindPrevious(
-      const orbit_client_protos::TimerInfo& from);
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindNext(
-      const orbit_client_protos::TimerInfo& from);
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindTop(
-      const orbit_client_protos::TimerInfo& from);
-  [[nodiscard]] const orbit_client_protos::TimerInfo* FindDown(
-      const orbit_client_protos::TimerInfo& from);
-  [[nodiscard]] std::pair<const orbit_client_protos::TimerInfo*,
-                          const orbit_client_protos::TimerInfo*>
-  GetMinMaxTimerInfoForFunction(uint64_t function_id) const;
 
   // TODO(http://b/194777907): Move GetColor outside TimeGraph
   [[nodiscard]] static Color GetColor(uint32_t id) {
@@ -161,23 +130,8 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
     return GetColor(static_cast<uint32_t>(tid));
   }
 
-  void SetIteratorOverlayData(
-      const absl::flat_hash_map<uint64_t, const orbit_client_protos::TimerInfo*>&
-          iterator_timer_info,
-      const absl::flat_hash_map<uint64_t, uint64_t>& iterator_id_to_function_id) {
-    iterator_timer_info_ = iterator_timer_info;
-    iterator_id_to_function_id_ = iterator_id_to_function_id;
-    RequestUpdate();
-  }
-
-  [[nodiscard]] float GetVerticalScrollingOffset() const { return vertical_scrolling_offset_; }
-  void SetVerticalScrollingOffset(float value);
-
   [[nodiscard]] uint64_t GetCaptureMin() const { return capture_min_timestamp_; }
   [[nodiscard]] uint64_t GetCaptureMax() const { return capture_max_timestamp_; }
-
-  [[nodiscard]] bool HasFrameTrack(uint64_t function_id) const;
-  void RemoveFrameTrack(uint64_t function_id);
 
   [[nodiscard]] std::vector<CaptureViewElement*> GetAllChildren() const override;
   [[nodiscard]] std::vector<CaptureViewElement*> GetNonHiddenChildren() const override;
@@ -189,10 +143,6 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
  protected:
   void PrepareBatcherAndUpdatePrimitives(PickingMode picking_mode);
   void DoUpdateLayout() override;
-  void DoDraw(Batcher& batcher, TextRenderer& text_renderer,
-              const DrawContext& draw_context) override;
-
-  void UpdateTracksPosition();
 
   [[nodiscard]] std::unique_ptr<orbit_accessibility::AccessibleInterface>
   CreateAccessibleInterface() override;
@@ -202,18 +152,8 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
   void ProcessPageFaultsTrackingTimer(const orbit_client_protos::TimerInfo& timer_info);
 
  private:
-  void DrawOverlay(Batcher& batcher, TextRenderer& text_renderer, PickingMode picking_mode);
-  void DrawIteratorBox(Batcher& batcher, TextRenderer& text_renderer, Vec2 pos, Vec2 size,
-                       const Color& color, const std::string& label, const std::string& time,
-                       float text_box_y);
-  void DrawIncompleteDataIntervals(Batcher& batcher, PickingMode picking_mode);
-
   AccessibleInterfaceProvider* accessible_parent_;
   TextRenderer text_renderer_static_;
-
-  // First member is id.
-  absl::flat_hash_map<uint64_t, const orbit_client_protos::TimerInfo*> iterator_timer_info_;
-  absl::flat_hash_map<uint64_t, uint64_t> iterator_id_to_function_id_;
 
   double ref_time_us_ = 0;
   double min_time_us_ = 0;
@@ -221,7 +161,6 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
   uint64_t capture_min_timestamp_ = std::numeric_limits<uint64_t>::max();
   uint64_t capture_max_timestamp_ = 0;
   double time_window_us_ = 0;
-  float vertical_scrolling_offset_ = 0;
 
   TimeGraphLayout layout_;
 
@@ -229,11 +168,10 @@ class TimeGraph final : public orbit_gl::CaptureViewElement,
 
   Batcher batcher_;
 
-  std::unique_ptr<orbit_gl::TrackManager> track_manager_;
+  std::unique_ptr<orbit_gl::TrackContainer> track_container_;
 
   ManualInstrumentationManager* manual_instrumentation_manager_;
   const orbit_client_data::CaptureData* capture_data_ = nullptr;
-  orbit_client_data::ThreadTrackDataProvider* thread_track_data_provider_ = nullptr;
 
   OrbitApp* app_ = nullptr;
 };
