+++ src/OrbitGl/TrackManager.cpp
@@ -30,6 +30,7 @@
 #include "OrbitBase/ThreadConstants.h"
 #include "TimeGraph.h"
 #include "TimeGraphLayout.h"
+#include "TrackContainer.h"
 #include "Viewport.h"
 
 using orbit_client_data::CallstackData;
@@ -37,13 +38,15 @@ using orbit_client_protos::TimerInfo;
 
 namespace orbit_gl {
 
-TrackManager::TrackManager(TimeGraph* time_graph, Viewport* viewport, TimeGraphLayout* layout,
-                           OrbitApp* app, orbit_client_data::CaptureData* capture_data)
-    : time_graph_(time_graph),
-      viewport_(viewport),
+TrackManager::TrackManager(TrackContainer* track_container, TimelineInfoInterface* timeline_info,
+                           Viewport* viewport, TimeGraphLayout* layout, OrbitApp* app,
+                           orbit_client_data::CaptureData* capture_data)
+    : viewport_(viewport),
       layout_(layout),
       capture_data_{capture_data},
-      app_{app} {
+      app_{app},
+      track_container_(track_container),
+      timeline_info_(timeline_info) {
   ORBIT_CHECK(capture_data != nullptr);
   for (Track::Type type : Track::kAllTrackTypes) {
     track_type_visibility_[type] = true;
@@ -209,8 +212,8 @@ void TrackManager::UpdateVisibleTrackList() {
       }
     }
   }
-  if (time_graph_ != nullptr) {
-    time_graph_->RequestUpdate();
+  if (track_container_ != nullptr) {
+    track_container_->RequestUpdate();
   }
 }
 
@@ -253,7 +256,7 @@ void TrackManager::UpdateMovingTrackPositionInVisibleTracks() {
 
   if (moving_track_previous_position != -1) {
     Track* moving_track = visible_tracks_[moving_track_previous_position];
-    time_graph_->VerticallyMoveIntoView(*moving_track);
+    track_container_->VerticallyMoveIntoView(*moving_track);
 
     visible_tracks_.erase(visible_tracks_.begin() + moving_track_previous_position);
 
@@ -354,8 +357,8 @@ void TrackManager::RemoveFrameTrack(uint64_t function_id) {
 
 void TrackManager::SetTrackTypeVisibility(Track::Type type, bool value) {
   track_type_visibility_[type] = value;
-  if (time_graph_ != nullptr) {
-    time_graph_->RequestUpdate();
+  if (track_container_ != nullptr) {
+    track_container_->RequestUpdate();
   }
   visible_track_list_needs_update_ = true;
 }
@@ -371,8 +374,8 @@ const absl::flat_hash_map<Track::Type, bool> TrackManager::GetAllTrackTypesVisib
 void TrackManager::RestoreAllTrackTypesVisibility(
     const absl::flat_hash_map<Track::Type, bool>& values) {
   track_type_visibility_ = values;
-  if (time_graph_ != nullptr) {
-    time_graph_->RequestUpdate();
+  if (track_container_ != nullptr) {
+    track_container_->RequestUpdate();
   }
   visible_track_list_needs_update_ = true;
 }
@@ -434,7 +437,7 @@ SchedulerTrack* TrackManager::GetOrCreateSchedulerTrack() {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   if (scheduler_track_ == nullptr) {
     auto [unused, timer_data] = capture_data_->CreateTimerData();
-    scheduler_track_ = std::make_shared<SchedulerTrack>(time_graph_, time_graph_, viewport_,
+    scheduler_track_ = std::make_shared<SchedulerTrack>(track_container_, timeline_info_, viewport_,
                                                         layout_, app_, capture_data_, timer_data);
     AddTrack(scheduler_track_);
   }
@@ -447,8 +450,8 @@ ThreadTrack* TrackManager::GetOrCreateThreadTrack(uint32_t tid) {
   if (track == nullptr) {
     auto thread_track_data_provider = capture_data_->GetThreadTrackDataProvider();
     thread_track_data_provider->CreateScopeTreeTimerData(tid);
-    track = std::make_shared<ThreadTrack>(time_graph_, time_graph_, viewport_, layout_, tid, app_,
-                                          capture_data_, thread_track_data_provider);
+    track = std::make_shared<ThreadTrack>(track_container_, timeline_info_, viewport_, layout_, tid,
+                                          app_, capture_data_, thread_track_data_provider);
     thread_tracks_[tid] = track;
     AddTrack(track);
   }
@@ -463,9 +466,9 @@ GpuTrack* TrackManager::GetOrCreateGpuTrack(uint64_t timeline_hash) {
   if (track == nullptr) {
     auto [unused1, submission_timer_data] = capture_data_->CreateTimerData();
     auto [unused2, marker_timer_data] = capture_data_->CreateTimerData();
-    track =
-        std::make_shared<GpuTrack>(time_graph_, time_graph_, viewport_, layout_, timeline_hash,
-                                   app_, capture_data_, submission_timer_data, marker_timer_data);
+    track = std::make_shared<GpuTrack>(track_container_, timeline_info_, viewport_, layout_,
+                                       timeline_hash, app_, capture_data_, submission_timer_data,
+                                       marker_timer_data);
     gpu_tracks_[timeline] = track;
     AddTrack(track);
   }
@@ -476,8 +479,8 @@ VariableTrack* TrackManager::GetOrCreateVariableTrack(const std::string& name) {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   std::shared_ptr<VariableTrack> track = variable_tracks_[name];
   if (track == nullptr) {
-    track = std::make_shared<VariableTrack>(time_graph_, time_graph_, viewport_, layout_, name,
-                                            capture_data_);
+    track = std::make_shared<VariableTrack>(track_container_, timeline_info_, viewport_, layout_,
+                                            name, capture_data_);
     variable_tracks_[name] = track;
     AddTrack(track);
   }
@@ -489,8 +492,8 @@ AsyncTrack* TrackManager::GetOrCreateAsyncTrack(const std::string& name) {
   std::shared_ptr<AsyncTrack> track = async_tracks_[name];
   if (track == nullptr) {
     auto [unused, timer_data] = capture_data_->CreateTimerData();
-    track = std::make_shared<AsyncTrack>(time_graph_, time_graph_, viewport_, layout_, name, app_,
-                                         capture_data_, timer_data);
+    track = std::make_shared<AsyncTrack>(track_container_, timeline_info_, viewport_, layout_, name,
+                                         app_, capture_data_, timer_data);
     async_tracks_[name] = track;
     AddTrack(track);
   }
@@ -507,8 +510,8 @@ FrameTrack* TrackManager::GetOrCreateFrameTrack(
   }
 
   auto [unused, timer_data] = capture_data_->CreateTimerData();
-  auto track = std::make_shared<FrameTrack>(time_graph_, time_graph_, viewport_, layout_, function,
-                                            app_, capture_data_, timer_data);
+  auto track = std::make_shared<FrameTrack>(track_container_, timeline_info_, viewport_, layout_,
+                                            function, app_, capture_data_, timer_data);
 
   // Normally we would call AddTrack(track) here, but frame tracks are removable by users
   // and therefore cannot be simply thrown into the flat vector of tracks. Also, we don't want to
@@ -522,8 +525,8 @@ FrameTrack* TrackManager::GetOrCreateFrameTrack(
 SystemMemoryTrack* TrackManager::CreateAndGetSystemMemoryTrack() {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   if (system_memory_track_ == nullptr) {
-    system_memory_track_ = std::make_shared<SystemMemoryTrack>(time_graph_, time_graph_, viewport_,
-                                                               layout_, capture_data_);
+    system_memory_track_ = std::make_shared<SystemMemoryTrack>(track_container_, timeline_info_,
+                                                               viewport_, layout_, capture_data_);
     AddTrack(system_memory_track_);
   }
 
@@ -535,7 +538,7 @@ CGroupAndProcessMemoryTrack* TrackManager::CreateAndGetCGroupAndProcessMemoryTra
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   if (cgroup_and_process_memory_track_ == nullptr) {
     cgroup_and_process_memory_track_ = std::make_shared<CGroupAndProcessMemoryTrack>(
-        time_graph_, time_graph_, viewport_, layout_, cgroup_name, capture_data_);
+        track_container_, timeline_info_, viewport_, layout_, cgroup_name, capture_data_);
     AddTrack(cgroup_and_process_memory_track_);
   }
 
@@ -547,8 +550,8 @@ PageFaultsTrack* TrackManager::CreateAndGetPageFaultsTrack(const std::string& cg
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   if (page_faults_track_ == nullptr) {
     page_faults_track_ =
-        std::make_shared<PageFaultsTrack>(time_graph_, time_graph_, viewport_, layout_, cgroup_name,
-                                          memory_sampling_period_ms, capture_data_);
+        std::make_shared<PageFaultsTrack>(track_container_, timeline_info_, viewport_, layout_,
+                                          cgroup_name, memory_sampling_period_ms, capture_data_);
     AddTrack(page_faults_track_);
   }
   return GetPageFaultsTrack();
